requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.1
Django==4.1.6
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
pycparser==2.21
PyJWT==1.7.1
pytz==2022.7.1
sqlparse==0.4.3
